% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.bridge.R
\name{cdcp.regression.bridge.plot}
\alias{cdcp.regression.bridge.plot}
\title{Plot of Monitoring Bridge}
\usage{
cdcp.regression.bridge.plot(bridge)
}
\arguments{
\item{bridge}{The output from running the \code{cdcp.regression.bridge(...)} function.}
}
\description{
Plot the so-called monitoring bridge using the output of the \code{cdcp.regression.bridge(...)} function.
}
\examples{

Example 1:
data <- cdcp.regression.data.sim()
bridge <- cdcp.regression.bridge(data$data, data$index_val)
cdcp.regression.bridge.plot(bridge)

Example 2: A simple example 
n <- 100
x <- 1:n
X <- cbind(1, x)
mu <- 3*(x <= 50) + x
y <- mu + rnorm(n)

# plot of simulated data
plot(x, y)
lines(x, mu)

# create list of data for CDCP Regression analysis 
index <- x
group <- x*0 + 1
data <- cdcp.regression.data(y = y, X = X, index = index, group = group)

# run monitoring bridge model and plot results
bridge <- cdcp.regression.bridge(data, index_val = 10:90)
cdcp.regression.bridge.plot(bridge)
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
