% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.bridge.R
\name{cdcp.regression.bridge}
\alias{cdcp.regression.bridge}
\title{Monitoring Bridge}
\usage{
cdcp.regression.bridge(data, index_val, cores = 4)
}
\arguments{
\item{data}{A CDCP Regression data object generated by the cdcp.regression.data(...) function.}

\item{index_val}{A sequence a sequence of indexes representing potential locations of chage points.}
}
\value{
a list consisting of two vectors:
\describe{
\item{index_val}{The same sequence of indexes as in the argument.}
\item{bbridge_val}{The computed monitoring bridge values.}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
This function compute the monitoring bridge for a set of indexes (index_val). It generate the necessary values to make a monitoring bridge plot for the cdcp.regression.bridge.plot(...) function.
}
\examples{
data <- cdcp.regression.data.sim()
model <- cdcp.regression.bridge(data$data, data$index_val)
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
