% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.data.R
\name{cdcp.regression.data.find.index.val}
\alias{cdcp.regression.data.find.index.val}
\title{Suggested Change Point Location}
\usage{
cdcp.regression.data.find.index.val(data, index_dummy = FALSE, threshold = 1)
}
\arguments{
\item{data}{A data object/list that specify the structure of the model created by the \code{cdcp.regression.data}.}

\item{index_dummy}{Should be \verb{= TRUE} if index dummies are used in the model; see \code{cdcp.regression.data} for additional details.}

\item{threshold}{Minimum number of effective data points per estimated parameter.}
}
\value{
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales.
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales...
An object from \code{cdcp.regression.estimation(...)} is a list containing the following components:
\describe{
\item{ln_max}{First item}
\item{aic}{Second item}
\item{Dn}{ddd}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
This function returns the largest set of indexes so that the model has at least \code{n} (specified by \code{threshold}) number of data points
to the left and right of all possible change points. This function is most useful for large panel data with several
dummy variables.
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
