% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.data.R
\name{cdcp.regression.data.sim}
\alias{cdcp.regression.data.sim}
\title{Simulate Test Data for the CDCPRegression Package}
\usage{
cdcp.regression.data.sim(seed = 1, beta = c(0.3, 0.35), sigma = 0.1)
}
\arguments{
\item{seed}{a seed for the random number generator.}

\item{beta}{the intercept to the left and right of the change point.}

\item{sigma}{the standard deviation used in simulating data.}
}
\value{
Output of \code{cdcp.regression.data(...)} for the simulated data.
}
\description{
Simulate data from the model \eqn{y_{i, j} = \beta_{\rm L} + (\beta_{\rm R} - \beta_{\rm L}) I(t_i > 1950) + \gamma z_i + \epsilon_{i, j}}, for
\eqn{t_i = 1900, \ldots , 2000}, \eqn{z_i = i/n}, \eqn{j = 1, \ldots m}, \eqn{n = 100} and \eqn{m = 4}, and with independent \eqn{\epsilon_{i, j} \sim {\rm N}(0, \sigma^2)}. The simulated data set is then aggregate this using \code{cdcp.regression.data(...)} function to create a toy data
set that can be used to test the functionality of the CDCPRegression package.
}
\examples{
data <- cdcp.regression.data.sim()
}
\references{
Hermansen, G., Knutsen, Carl Henrik & Nygaard, Haavard Mokleiv. (2021). Characterizing and assessing temporal heterogeneity: Introducing a change point framework, with applications on the study of democratization, Political Analysis, 29, 485-504

Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
