% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.estimation.R
\name{cdcp.regression.estimate}
\alias{cdcp.regression.estimate}
\title{Estimating the parameters in the model}
\usage{
cdcp.regression.estimate(data, index_val)
}
\arguments{
\item{data}{A data object/list that specify the structure of the model created by the}

\item{index_val}{Vector of indexes representing the locations for a potential change point}
}
\value{
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales.
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales...
An object from cdcp.regression.estimation(...) is a list containing the following components:
\describe{
\item{ln_max}{First item}
\item{aic}{Second item}
\item{Dn}{ddd}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
This function estimate the parameters, including the location of the change point,
for the specified model. The model is specified through a data object/list, which
should be created using the cdcp.regression.data(...) function.
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
