% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.cdcp.R
\name{cdcp.regression.plot}
\alias{cdcp.regression.plot}
\title{Plot of Confidence Sets for the Location of a Change Point}
\usage{
cdcp.regression.plot(cd, half_correction = TRUE)
}
\arguments{
\item{half_correction}{if \code{TRUE} the confidence sets are plotted with half correction.}

\item{cdcp}{the output of the \code{cdcp.regression(...)} function.}
}
\description{
Plot the confidence sets for the location of a change point for panel regression. The
plotting function is quite minimal, however, it should be easy to modify this to individual needs.
}
\examples{
data <- cdcp.regression.data.sim()
model <- cdcp.regression.estimate(data$data, data$index_val)
cd <- cdcp.regression(data$data, model, data$index_val)
cdcp.regression.plot(cd)
}
\references{
Hermansen, G., Knutsen, Carl Henrik & Nygaard, Haavard Mokleiv. (2021). Characterizing and assessing temporal heterogeneity: Introducing a change point framework, with applications on the study of democratization, Political Analysis, 29, 485-504

Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
