% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.cdcp.R
\name{cdcp.regression}
\alias{cdcp.regression}
\title{Confidence Sets for the Location of a Change Point}
\usage{
cdcp.regression(
  data,
  model,
  index_val,
  boot = 100,
  index_val_sub = NULL,
  u = seq(0, 1, length.out = 100),
  boot_type = c("gaussian", "independent", "group"),
  cores = 4
)
}
\arguments{
\item{data}{the output from the \code{cdcp.regression.data(...)} function.}

\item{model}{the output of the \code{cdcp.regression.estimation(...)} function.}

\item{index_val}{vector of indexes representing potential locations for a change point.}

\item{boot}{the number of bootstrap samples.}

\item{index_val_sub}{a subset of \code{index_val} that can be used to speed up calculations if the approximate location of the change point is known.}

\item{u}{the resolution used to compute the confidence sets (a sequence of numbers between 0 and 1).}

\item{boot_type}{the method used to sample residuals for the bootstrap: \code{gaussian} use use parametric
bootstrap under the assumption of a Gaussian model, \code{independent} makes independent draws from the
estimated residuals and \code{group} use the assumed group structure to draw independent residuals.}
}
\value{
A list containing the following:
\describe{
\item{cc_set}{a matrix representation of the confidense set for all levels u and index_val.}
\item{cc_set_half}{same as above but with half correction.}
\item{index_val}{same as \code{index_val} argument}
\item{u}{same as \code{u} argument}
}
}
\description{
This function computes the confidence sets, using bootstrapping, for the location of a change point; see Hermansen (2021) for details.
}
\examples{
data <- cdcp.regression.data.sim()
model <- cdcp.regression.estimate(data$data, data$index_val)
cd <- cdcp.regression(data$data, model, data$index_val)
}
\references{
Hermansen, G., Knutsen, Carl Henrik & Nygaard, Haavard Mokleiv. (2021). Characterizing and assessing temporal heterogeneity: Introducing a change point framework, with applications on the study of democratization, Political Analysis, 29, 485-504

Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
