% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.doc.R
\name{cdcp.regression.beta.doc.plot}
\alias{cdcp.regression.beta.doc.plot}
\title{Plot Confidence Curve for the Degree of Change}
\usage{
cdcp.regression.beta.doc.plot(doc, approx = TRUE)
}
\arguments{
\item{doc}{the output from running the \code{cdcp.regression.data(...)} function.}

\item{approx}{should the approximation of the confidence curve be plotted (\code{approx = TRUE} will plot this).}
}
\description{
Plot the confidence curve (cc) for the degree of change.
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
