% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.doc.R
\name{cdcp.regression.beta.doc}
\alias{cdcp.regression.beta.doc}
\title{Confidence Distribution for the Degree of Change}
\usage{
cdcp.regression.beta.doc(model, k, delta_val, index_val, boot = 0, boot_type)
}
\arguments{
\item{model}{The estimated model object from \code{cdcp.regression.estimation(...)}.}

\item{index_val}{A consecutive sequence of indexes representing the locations for a potential change point.}

\item{boot}{The number of bootstrap samples.}

\item{boot_type}{A subset of \code{index_val} used to speed up the calculations.}

\item{data}{A data object/list that specify the structure of the model created by the \code{cdcp.regression.data(...)}.}
}
\value{
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales.
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales...
An object from \code{cdcp.regression.estimation(...)} is a list containing the following components:
\describe{
\item{ln_max}{First item}
\item{aic}{Second item}
\item{Dn}{ddd}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
This function can compute both the bootstrap and the approximation...
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
