% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.bridge.R
\name{cdcp.regression.bridge.plot}
\alias{cdcp.regression.bridge.plot}
\title{Plot of Monitoring Bridge}
\usage{
cdcp.regression.bridge.plot(bridge)
}
\arguments{
\item{bridge}{The output from running the \code{cdcp.regression.bridge(...)} function.}
}
\description{
Plot the so-called monitoring bridge using the output of the \code{cdcp.regression.bridge(...)} function.
}
\examples{
data <- cdcp.regression.data.sim()
bridge <- cdcp.regression.bridge(data$data, data$index_val)
cdcp.regression.bridge.plot(bridge)
}
\references{
Hermansen, G., Knutsen, Carl Henrik & Nygaard, Haavard Mokleiv. (2021). Characterizing and assessing temporal heterogeneity: Introducing a change point framework, with applications on the study of democratization, Political Analysis, 29, 485-504

Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
