% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.bridge.R
\name{cdcp.regression.bridge}
\alias{cdcp.regression.bridge}
\title{Monitoring Bridge}
\usage{
cdcp.regression.bridge(data, index_val, cores = 4)
}
\arguments{
\item{data}{the output from the \code{cdcp.regression.data(...)} function.}

\item{index_val}{vector of indexes representing potential locations for a change point.}
}
\value{
a list consisting of two vectors:
\describe{
\item{index_val}{same as \code{index_val} argument}
\item{bbridge_val}{the computed monitoring bridge values.}
}
}
\description{
This function compute the monitoring bridge for a set of indexes (index_val); see Hermansen (2021) for details.
}
\examples{
data <- cdcp.regression.data.sim()
model <- cdcp.regression.bridge(data$data, data$index_val)
}
\references{
Hermansen, G., Knutsen, Carl Henrik & Nygaard, Haavard Mokleiv. (2021). Characterizing and assessing temporal heterogeneity: Introducing a change point framework, with applications on the study of democratization, Political Analysis, 29, 485-504

Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
