% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.data.R
\name{cdcp.regression.data}
\alias{cdcp.regression.data}
\title{Panel and Regression Data for Change Point}
\usage{
cdcp.regression.data(
  y,
  X,
  Z = NULL,
  D = NULL,
  index,
  group,
  index_dummy = FALSE,
  group_dummy = FALSE,
  lag = FALSE
)
}
\arguments{
\item{y}{The response or dependent variable.}

\item{X}{The covariates where we will examine the possibility of a change point. Categorical variables has to be cnverted before...
Note that the model is constructed such that there will be looking for change point among all variables in X.}

\item{Z}{Protected covariates, meaning covariates that we beilve are constant.}

\item{D}{Protected categorical covariates.}

\item{index}{List of idexes (e.g. years) that}

\item{group}{A list of}

\item{index_dummy}{Set \code{index_dummy = TRUE} to incude a categorical variable for each index (e.g. a year dummy) in the model.}

\item{group_dummy}{Set \code{group_dummy = TRUE} to include group specific dummy varaibles.}

\item{lag}{Set \code{lag = TRUE} to include the lagged respose \code{y} as a covariate in the model.}
}
\value{
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales.
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales...
An object from \code{cdcp.regression.estimation(...)} is a list containing the following components:
\describe{
\item{ln_max}{First item}
\item{aic}{Second item}
\item{Dn}{ddd}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
This function compiles data and can also be used to add common dummy variables.
}
\examples{
y <- rnorm(100)
mu <- 0.5*x + 0.1*5
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
