% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcp.regression.cdcp.R
\name{cdcp.regression}
\alias{cdcp.regression}
\title{Confidence Distribution for the Location of the Change Point}
\usage{
cdcp.regression(
  data,
  model,
  index_val,
  boot = 100,
  index_val_sub = NULL,
  u = seq(0, 1, length.out = 100),
  boot_type = c("gaussian", "independent", "group")
)
}
\arguments{
\item{data}{A data object/list that specify the structure of the model created by the \code{cdcp.regression.data(...)}.}

\item{model}{The estimated model object from \code{cdcp.regression.estimation(...)}.}

\item{index_val}{A consecutive sequence of indexes representing the locations for a potential change point.}

\item{boot}{The number of bootstrap samples.}

\item{index_val_sub}{A subset of \code{index_val} used to speed up the calculations.}

\item{u}{The resolution for the probability...}

\item{boot_type}{The method used to sample residuales for the bootstrap \verb{= "group"} ...}
}
\value{
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales.
A list of estimated parameters, the maximised log-likelihood, aic, deviance and residuales...
An object from \code{cdcp.regression.estimation(...)} is a list containing the following components:
\describe{
\item{ln_max}{First item}
\item{aic}{Second item}
\item{Dn}{ddd}
}
\itemize{
\item{"parameter 1"}{Stuff}
\item{}{Stuff}
}
}
\description{
\code{cdcp.regression(...)} compute a bootstrap for the change point
}
\references{
Cunen, C., Hermansen, G., & Hjort, N. L. (2018). Confidence distributions for change-points and regime shifts. Journal of Statistical Planning and Inference, 195, 14-34.
}
